pytorch-deploy/requirements1.txt
-f https://download.pytorch.org/whl/torch_stable.html
appdirs==1.4.4
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.1.3
colorama==0.4.4
cssselect==1.1.0
distlib==0.3.4
et-xmlfile==1.1.0
fake-useragent==0.1.11
filelock==3.4.0
Flask==2.2.2
gevent==21.12.0
graphviz==0.20
greenlet==1.1.2
grequests==0.6.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.7.1
MarkupSafe==2.1.1
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
parse==1.19.0
Pillow==9.3.0
pipenv==2021.11.23
platformdirs==2.4.0
plotly==5.9.0
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
pyppeteer==0.2.6
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.26.0
requests-html==0.10.0
responses==0.16.0
six==1.16.0
soupsieve==2.3.1
tenacity==8.0.1
toml==0.10.2
torch==1.13.1+cpu
torchvision==0.14.1+cpu
tqdm==4.62.3
typing_extensions==4.4.0
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
w3lib==1.22.0
websockets==9.1
Werkzeug==2.2.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
